refactor(types): Separate Types.ts into files
This commit splits the original Types.ts into dedicated files under
the types/ directory. An index.ts re-exports them for simpler
maintenance and improved clarity.
Tests are updated to reflect the changes.
