+++ src/index.ts
@@ -1,2 +1,1 @@
-export * from "./Types";
 export * from "./Workflow";
