+++ src/types/Common.ts
@@ -1,0 +1,16 @@
+/*
+ * This file stores cross-category or shared types.
+ *
+ * These types could logically belong to multiple categories,
+ * such as both Utility and Context. By placing them in a single file,
+ * we avoid arbitrary classification and maintain a clear separation
+ * of responsibilities, enabling simpler management and reuse.
+ *
+ * Copyright (c) 2015-2025 Yuba Technology. All rights reserved.
+ * This file is a collaborative effort of the Yuba Technology team
+ * and all contributors to the WorkflowX project.
+ *
+ * Licensed under the AGPLv3 license.
+ */
+
+// import { Step, StepReturnType } from "./Step";
